Cleanup Windows window code
