+++ src/utilities/windows/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(NOX_UTILITIES_WINDOWS_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/CMakeLists.txt
+    ${CMAKE_CURRENT_SOURCE_DIR}/windows_helper.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/windows_helper.h
+)
+
+target_sources(${NOX_LIB_NAME} PRIVATE ${NOX_UTILITIES_WINDOWS_SOURCES})
