+++ src/utilities/windows/windows_helper.h
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <nox/export.h>
+#include <nox/maths/vector2d.h>
+
+#include <Windows.h>
+
+#include <string_view>
+
+namespace NOX {
+
+struct WindowDescriptor;
+
+namespace WindowsHelper {
+
+NOX_EXPORT void registerWindowClass(const WNDCLASS &attributes);
+NOX_EXPORT void unregisterWindowClass(std::string_view className);
+
+DWORD getWindowStyle(const WindowDescriptor &descriptor);
+RECT getWindowClientArea(const WindowDescriptor &descriptor, DWORD style);
+Vector2D<int32_t> getWindowPosition(const WindowDescriptor &descriptor, RECT clientArea);
+Vector2D<uint32_t> getWindowSize(RECT clientArea);
+
+} // namespace WindowsHelper
+
+} // namespace NOX
