--- src/window/windows/windows_window_class.h
@@ -1,13 +1,0 @@
-#pragma once
-
-namespace NOX {
-
-class WindowsWindowClass {
-  public:
-    static constexpr auto className = "__NOX_WINDOWS_WINDOW_CLASS__";
-
-    WindowsWindowClass();
-    ~WindowsWindowClass();
-};
-
-} // namespace NOX
