--- src/window/windows/windows_window_procedure.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <Windows.h>
-
-namespace NOX {
-
-void populateWindowMessageHandlers();
-LRESULT CALLBACK windowProcedure(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
-
-} // namespace NOX
